src/requirements.txt
mcp==1.6.0
uvicorn==0.34.0
pyyaml
requests